requirements.txt
FastAPI
jinja2
SQLAlchemy
uvicorn
setuptools_scm
toml
pyyaml
git+https://github.com/Tina-otoge/default-sqlalchemy-setup
